move manifest/hook partition code out of tiller and export
