+++ pkg/tiller/kind_sorter.go
@@ -17,142 +17,19 @@ limitations under the License.
 package tiller
 
 import (
-	"sort"
+	"k8s.io/helm/pkg/manifest"
 )
 
-// SortOrder is an ordering of Kinds.
-type SortOrder []string
+// SortOrder has been aliased to avoid breaking API
+type SortOrder = manifest.SortOrder
 
-// InstallOrder is the order in which manifests should be installed (by Kind).
-//
-// Those occurring earlier in the list get installed before those occurring later in the list.
-var InstallOrder SortOrder = []string{
-	"Namespace",
-	"ResourceQuota",
-	"LimitRange",
-	"PodSecurityPolicy",
-	"PodDisruptionBudget",
-	"Secret",
-	"ConfigMap",
-	"StorageClass",
-	"PersistentVolume",
-	"PersistentVolumeClaim",
-	"ServiceAccount",
-	"CustomResourceDefinition",
-	"ClusterRole",
-	"ClusterRoleBinding",
-	"Role",
-	"RoleBinding",
-	"Service",
-	"DaemonSet",
-	"Pod",
-	"ReplicationController",
-	"ReplicaSet",
-	"Deployment",
-	"StatefulSet",
-	"Job",
-	"CronJob",
-	"Ingress",
-	"APIService",
-}
+var (
+	// InstallOrder has been aliased to avoid breaking API
+	InstallOrder SortOrder = manifest.InstallOrder
 
-// UninstallOrder is the order in which manifests should be uninstalled (by Kind).
-//
-// Those occurring earlier in the list get uninstalled before those occurring later in the list.
-var UninstallOrder SortOrder = []string{
-	"APIService",
-	"Ingress",
-	"Service",
-	"CronJob",
-	"Job",
-	"StatefulSet",
-	"Deployment",
-	"ReplicaSet",
-	"ReplicationController",
-	"Pod",
-	"DaemonSet",
-	"RoleBinding",
-	"Role",
-	"ClusterRoleBinding",
-	"ClusterRole",
-	"CustomResourceDefinition",
-	"ServiceAccount",
-	"PersistentVolumeClaim",
-	"PersistentVolume",
-	"StorageClass",
-	"ConfigMap",
-	"Secret",
-	"PodDisruptionBudget",
-	"PodSecurityPolicy",
-	"LimitRange",
-	"ResourceQuota",
-	"Namespace",
-}
+	// UninstallOrder has been aliased to avoid breaking API
+	UninstallOrder SortOrder = manifest.UninstallOrder
 
-// sortByKind does an in-place sort of manifests by Kind.
-//
-// Results are sorted by 'ordering'
-func sortByKind(manifests []Manifest, ordering SortOrder) []Manifest {
-	ks := newKindSorter(manifests, ordering)
-	sort.Sort(ks)
-	return ks.manifests
-}
-
-type kindSorter struct {
-	ordering  map[string]int
-	manifests []Manifest
-}
-
-func newKindSorter(m []Manifest, s SortOrder) *kindSorter {
-	o := make(map[string]int, len(s))
-	for v, k := range s {
-		o[k] = v
-	}
-
-	return &kindSorter{
-		manifests: m,
-		ordering:  o,
-	}
-}
-
-func (k *kindSorter) Len() int { return len(k.manifests) }
-
-func (k *kindSorter) Swap(i, j int) { k.manifests[i], k.manifests[j] = k.manifests[j], k.manifests[i] }
-
-func (k *kindSorter) Less(i, j int) bool {
-	a := k.manifests[i]
-	b := k.manifests[j]
-	first, aok := k.ordering[a.Head.Kind]
-	second, bok := k.ordering[b.Head.Kind]
-
-	if !aok && !bok {
-		// if both are unknown then sort alphabetically by kind and name
-		if a.Head.Kind != b.Head.Kind {
-			return a.Head.Kind < b.Head.Kind
-		}
-		return a.Name < b.Name
-	}
-
-	// unknown kind is last
-	if !aok {
-		return false
-	}
-	if !bok {
-		return true
-	}
-
-	// if same kind sub sort alphanumeric
-	if first == second {
-		return a.Name < b.Name
-	}
-	// sort different kinds
-	return first < second
-}
-
-// SortByKind sorts manifests in InstallOrder
-func SortByKind(manifests []Manifest) []Manifest {
-	ordering := InstallOrder
-	ks := newKindSorter(manifests, ordering)
-	sort.Sort(ks)
-	return ks.manifests
-}
+	// SortByKind has been aliased to avoid breaking API
+	SortByKind func([]manifest.Manifest) []manifest.Manifest = manifest.SortByKind
+)
